+++ androidApp/src/main/kotlin/com/jermey/seal/SealApplication.kt
@@ -1,7 +1,7 @@
 package com.jermey.seal
 
 import android.app.Application
-import com.jermey.seal.android.ConscryptInitializer
+import com.jermey.seal.jvm.ConscryptInitializer
 import com.jermey.seal.demo.di.appModule
 import com.jermey.seal.demo.di.containerModule
 import org.koin.core.context.startKoin